+++ CHANGELOG.md
@@ -109,6 +109,7 @@
 84. Fixed Issue with Mods not running on older Unity Versions.
 85. Added Load Balancing Hosting for Assembly Generator RemoteAPI.    (Credits to DubyaDude and Samboy :D)
 86. Fixed Small Bug in the RemoteAPI Request System for the Assembly Generator.
+87. Added Compatibility Layer Resolver to MelonHandler for Future Changes.
 
 ---